HREOS-4416, corrección para que los campos acepten decimales

+++ recovery-webservice/src/main/java/es/pfsgroup/plugin/rem/model/GastoPrinex.java
@@ -205,22 +205,22 @@ public class GastoPrinex implements Serializable, Auditable {
 	private String caracteristica;
 	
 	@Column(name = "GPL_DIARIO1_BASE")
-	private Long diario1Base;
+	private Double diario1Base;
 	
 	@Column(name = "GPL_DIARIO1_TIPO")
-	private Long diario1Tipo;
+	private Double diario1Tipo;
 	
 	@Column(name = "GPL_DIARIO1_CUOTA")
-	private Long diario1Cuota;
+	private Double diario1Cuota;
 	
 	@Column(name = "GPL_DIARIO2_BASE")
-	private Long diario2Base;
+	private Double diario2Base;
 	
 	@Column(name = "GPL_DIARIO2_TIPO")
-	private Long diario2Tipo;
+	private Double diario2Tipo;
 	
 	@Column(name = "GPL_DIARIO2_CUOTA")
-	private Long diario2Cuota;
+	private Double diario2Cuota;
 	
 	
 	@Embedded
@@ -698,51 +698,51 @@ public class GastoPrinex implements Serializable, Auditable {
 		this.auditoria = auditoria;
 	}
 
-	public Long getDiario1Base() {
+	public Double getDiario1Base() {
 		return diario1Base;
 	}
 
-	public void setDiario1Base(Long diario1Base) {
+	public void setDiario1Base(Double diario1Base) {
 		this.diario1Base = diario1Base;
 	}
 
-	public Long getDiario1Tipo() {
+	public Double getDiario1Tipo() {
 		return diario1Tipo;
 	}
 
-	public void setDiario1Tipo(Long diario1Tipo) {
+	public void setDiario1Tipo(Double diario1Tipo) {
 		this.diario1Tipo = diario1Tipo;
 	}
 
-	public Long getDiario1Cuota() {
+	public Double getDiario1Cuota() {
 		return diario1Cuota;
 	}
 
-	public void setDiario1Cuota(Long diario1Cuota) {
+	public void setDiario1Cuota(Double diario1Cuota) {
 		this.diario1Cuota = diario1Cuota;
 	}
 
-	public Long getDiario2Base() {
+	public Double getDiario2Base() {
 		return diario2Base;
 	}
 
-	public void setDiario2Base(Long diario2Base) {
+	public void setDiario2Base(Double diario2Base) {
 		this.diario2Base = diario2Base;
 	}
 
-	public Long getDiario2Tipo() {
+	public Double getDiario2Tipo() {
 		return diario2Tipo;
 	}
 
-	public void setDiario2Tipo(Long diario2Tipo) {
+	public void setDiario2Tipo(Double diario2Tipo) {
 		this.diario2Tipo = diario2Tipo;
 	}
 
-	public Long getDiario2Cuota() {
+	public Double getDiario2Cuota() {
 		return diario2Cuota;
 	}
 
-	public void setDiario2Cuota(Long diario2Cuota) {
+	public void setDiario2Cuota(Double diario2Cuota) {
 		this.diario2Cuota = diario2Cuota;
 	}
